requirements.txt
autoattack @ git+https://github.com/fra31/auto-attack.git@a39220048b3c9f2cca9a4d3a54604793c68eca7e
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
filelock==3.14.0
fsspec==2024.5.0
gdown==5.1.0
geotorch==0.3.0
huggingface-hub==0.23.0
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.0
pandas==2.0.3
pillow==10.3.0
prettytable==3.10.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
robustbench @ git+https://github.com/RobustBench/robustbench.git@391406affed13d824daf4f228fd50dfe619684be
safetensors==0.4.3
scipy==1.10.1
setproctitle==1.3.3
six==1.16.0
soupsieve==2.5
sympy==1.12
timm==1.0.3
torch==2.3.0
torchdiffeq==0.2.3
torchvision==0.18.0
tqdm==4.66.4
triton==2.3.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yacs==0.1.8
